refactor(telebot): handle input for dynamic strategy - create TeleBotHandleInputDriverInterface and facade and default driver
