+++ new_structure/sys-modules/telebot/Lib/Drivers/TeleBotHandleInputDriverInterface.go
@@ -1,0 +1,7 @@
+package Drivers
+
+import tele "gopkg.in/telebot.v4"
+
+type TeleBotHandleInputDriverInterface interface {
+	HandleInput(isBtn bool, value string, c tele.Context) error
+}
